requirements.txt
asgiref==3.5.0
azure-core==1.24.0
azure-storage-blob==12.12.0
backports.zoneinfo==0.2.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-writable-nested==0.6.3
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
Markdown==3.3.6
msrest==0.6.21
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sentry-sdk==1.5.11
six==1.16.0
sqlparse==0.4.2
style==1.1.0
typing_extensions==4.2.0
update==0.0.1
urllib3==1.26.9
webencodings==0.5.1
zipp==3.7.0
